I hereby agree to the terms of the CLA available at: https://datafuse.rs/policies/cla/

## Summary
Sorry, pull  error, re apply ,inspect BooleanArray exit true 

 **Run sql:** 
`SELECT number FROM numbers_mt(100000000000) where number = 100 limit 1 ;`
     
**before:**
   1 rows in set. Elapsed: 66.156 sec. Processed 100.00 billion rows, 800.00 GB (1.51 billion rows/s., 12.09 GB/s.)
**after:**
   1 rows in set. Elapsed: 37.384 sec. Processed 100.00 billion rows, 800.00 GB (2.67 billion rows/s., 21.40 GB/s.)

## Changelog
- Performance Improvement