@sundy-li 2021-08-08
LGTM, did you benchmark other queries which not match this rule?

Like

```
select sum(number) from numbers_mt(10000000) where number != 100;
```