@ives9638 2021-08-08
> LGTM, did you benchmark other queries which not match this rule?
> 
> Like
> 
> ```
> select sum(number) from numbers_mt(10000000) where number != 100;
> ```

**1 select sum(number) from numbers_mt(10000000) where number != 100**
┌────sum(number)─┐
│ 49999994999900 │
└────────────────┘
1 rows in set. Elapsed: 0.014 sec. Processed 10.00 million rows, 80.00 MB (698.53 million rows/s., 5.59 GB/s.)
 
 **2 select sum(number) from numbers_mt(10000000)** 
┌────sum(number)─┐
│ 49999995000000 │
└────────────────┘
1 rows in set. Elapsed: 0.008 sec. Processed 10.00 million rows, 80.00 MB (1.23 billion rows/s., 9.88 GB/s.)

It should be the same as before optimization. Let me test:
Slow is:
arrow2::compute::filter::filter_ nonnull_ Primitive and arrow2:: compute:: comparison:: primitive:: NEQ_ scalar
 
![image](https://user-images.githubusercontent.com/83902299/128628219-53147fa5-4140-4e06-870d-720f6e07261a.png)